Transfered local work to repo
This commit is of changes I've made locally that I now transfered to the repo.

These changes contain new files JobApplication, JobApplicationController and JobApplicationRepository. The goal with them was to allow a user to apply for a job.

Changes were made to the PersonController to ask for personal info when registering a new account. The details requested correspond to what's in the database.

Added getters and setters to the Person class.

PersonService has now some added authentication by checking all fields are filled in when creating an account and that a valid email is being used by checking that an "@" is there.