Added JobApplication-files
Added all the JobApplication files. These are supposed to contain the relevant information a person applying for a job should enter. Once they're done with their application, it should be saved in the database.

From the admin's perspective, they should be allowed to search for and view applications and hire/reject it.

Currently, they don't fill any function. This was discussed during seminar 2 that since these cover 5.3 of the application-description and we already have fulfilled point 5.1, this was no longer needed.